Master to Nicolas

+++ src/main/java/MySpotLibrary/Entites/Player.java
@@ -14,7 +14,7 @@ public class Player {
 
     @Id
     @GeneratedValue
-    private int id;
+    private long id;
     private String username;
     private int level;
 
@@ -50,11 +50,11 @@ public class Player {
         this.level = level;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -129,4 +129,5 @@ public class Player {
     public void setMarks(List<Marking> marks) {
         this.marks = marks;
     }
+
 }
